cleaning up hg19 locus view code in main file
